requirements.txt
google-api-python-client==1.10.0
google-auth-httplib2==0.0.4
google-cloud-bigquery
pyarrow
pandas